🛠️refactor: edited export sentence in UploadImage
