+++ src/components/client/UploadImage/components/index.ts
@@ -1,0 +1,1 @@
+export { default as UploadImageItem } from './UploadImageItem/UploadImageItem';
