+++ src/components/client/UploadImage/context/index.ts
@@ -1,0 +1,1 @@
+export * from './UploadImageContext/UploadImageContext';
